% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmark_causative}
\alias{benchmark_causative}
\title{Benchmarking causative task}
\usage{
benchmark_causative(embedding_df, sig_level = 0.05, bootstraps = 10000,
  verbose = TRUE)
}
\arguments{
\item{embedding_df}{The embedding data frame with bound semantic type}

\item{sig_level}{The significance level threshold}

\item{bootstraps}{The number of boostraps to perform}

\item{verbose}{Flag for verbosity}
}
\value{
Dataframe of scores on this task per causative relationship
}
\description{
The benchmarking strategy leverages previously published ‘known’ relationships between medicalconcepts.
We compare how similar the embeddings for a pair of concepts are by computing the
cosine similarity of their corresponding vectors,
and use this similarity to assess whether or not thetwo concepts are related.
\code{\link{benchmark_causative}} assesses an embedding's ability to recover causes from
the UMLS' table (MRREL) of entities known to be the cause of a certain result.
}
