% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmark_comorbidities}
\alias{benchmark_comorbidities}
\title{Benchmarking comorbidities task}
\usage{
benchmark_comorbidities(embedding_df, sig_level = 0.05,
  bootstraps = 10000)
}
\arguments{
\item{embedding_df}{The embedding dataframe with bound semantic types}

\item{sig_level}{The significance level threshold}

\item{bootstraps}{The number of bootstraps to perform}
}
\value{
Dataframe of performance on this benchmark per disease
}
\description{
The benchmarking strategy leverages previously published ‘known’ relationships between medicalconcepts.
We compare how similar the embeddings for a pair of concepts are by computing the
cosine similarity of their corresponding vectors,
and use this similarity to assess whether or not thetwo concepts are related.
\code{\link{benchmark_comorbidities}} focuses on an embedding's ability to identify comorbidities.
A comorbidity is a disease or condition that frequently accompanies a  primary  diagnosis.
}
