% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmark_ndf_rt}
\alias{benchmark_ndf_rt}
\title{Benchmarking NDF RT task}
\usage{
benchmark_ndf_rt(embedding_df, sig_level = 0.05, bootstraps = 10000)
}
\arguments{
\item{embedding_df}{The embedding data frame with bound semantic type}

\item{sig_level}{The significance level threshold}

\item{bootstraps}{The number of bootstraps to perform}
}
\value{
Dataframe of performance on this task
}
\description{
The benchmarking strategy leverages previously published ‘known’ relationships between medicalconcepts.
We compare how similar the embeddings for a pair of concepts are by computing the
cosine similarity of their corresponding vectors,
and use this similarity to assess whether or not thetwo concepts are related.
\code{\link{benchmark_ndf_rt}} assesses an embedding's ability to power to detect "may treat" and "may prevent"
relationships using bootstrap scores of random drug-disease pairs.
}
