% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmark_semantic_type}
\alias{benchmark_semantic_type}
\title{Benchmarking semnatic type task}
\usage{
benchmark_semantic_type(embedding_df, sig_level = 0.05,
  bootstraps = 10000)
}
\arguments{
\item{embedding_df}{The embedding dataframe with bound semantic types}

\item{sig_level}{The significance level threshold}

\item{bootstraps}{The number of bootstraps to perform}
}
\value{
Dataframe of performance on this task per semantic type
}
\description{
The benchmarking strategy leverages previously published ‘known’ relationships between medicalconcepts.
We compare how similar the embeddings for a pair of concepts are by computing the
cosine similarity of their corresponding vectors,
and use this similarity to assess whether or not thetwo concepts are related.
\code{\link{benchmark_semantic_type}} assesses an ability to identify semantic types. Semantic types are
meta-information about which category a concept belongs to, and these categories are arranged in a hierarchy.
}
