% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmark_similarity}
\alias{benchmark_similarity}
\title{Benchmarking similarity task}
\usage{
benchmark_similarity(embedding_df)
}
\arguments{
\item{embedding_df}{The embedding dataframe}
}
\value{
Dataframe of relatedness
}
\description{
The benchmarking strategy leverages previously published ‘known’ relationships between medicalconcepts.
We compare how similar the embeddings for a pair of concepts are by computing the
cosine similarity of their corresponding vectors,
and use this similarity to assess whether or not thetwo concepts are related.
\code{link{benchmark_similarity}} reports the spearman correlation between the human assessment scores
and cosine similarity from the embeddings.
}
