% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_semantic_types}
\alias{bind_semantic_types}
\title{Bind Semantic Types}
\usage{
bind_semantic_types(embedding_df)
}
\arguments{
\item{embedding_df}{The embedding dataframe}
}
\value{
Embedding dataframe with CUI, SemanticType, and String columns prepended
}
\description{
Bind Semantic Types
}
