% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bootstrap_samples}
\alias{bootstrap_samples}
\title{Bootstrap samples}
\usage{
bootstrap_samples(query_db, results_db, bootstraps = 10000,
  eps = 1e-06)
}
\arguments{
\item{query_db}{Query database}

\item{results_db}{Results database}

\item{bootstraps}{Number of bootstraps to perform}

\item{eps}{Value to add to denominator to prevent division by zero}
}
\value{
Bootstrapped scores
}
\description{
Bootstrap samples
}
