% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_embedding_semantic_columns}
\alias{check_embedding_semantic_columns}
\title{Checking embedding semantic type columns}
\usage{
check_embedding_semantic_columns(embedding_df)
}
\arguments{
\item{embedding_df}{The embedding dataframe}
}
\value{
True/False that the first 3 columns are CUI, SemanticType, String
}
\description{
Checking embedding semantic type columns
}
