% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_paired_comparison.R
\name{compare_embeddings}
\alias{compare_embeddings}
\title{Run paired comparisons}
\usage{
compare_embeddings(embedding1_df, embedding2_df, benchmark = c("all",
  "comorbidities", "causitive", "ndfrt", "semantic_type", "similarity"))
}
\arguments{
\item{embedding1_df}{The first embedding's dataframe with bound semantic type}

\item{embedding2_df}{The second embedding's dataframe with bound semantic type}

\item{benchmark}{Which benchmarks to perform. Valid inputs are c("all", "comorbidities", "causative", "ndfrt", "semantic_type", "similarity")}
}
\value{
A list with two entries of result dataframes, one for each embedding.
}
\description{
Run paired comparisons
}
