% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glove_fit.R
\name{construct_glove_embedding}
\alias{construct_glove_embedding}
\title{Constructs GloVe embeddings from co-occurrence matrix}
\usage{
construct_glove_embedding(coccur, dim_size = 100, iters = 25)
}
\arguments{
\item{coccur}{The co-occurrence matrix}

\item{dim_size}{The embedding dimension}

\item{iters}{The number of iterations}
}
\value{
GloVe embedding
}
\description{
Constructs GloVe embeddings from co-occurrence matrix
}
