% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_fit.R
\name{construct_pca_embedding}
\alias{construct_pca_embedding}
\title{Constructs PCA embedding based raw co-occurrence counts}
\usage{
construct_pca_embedding(coccur, dim_size = 100, iters = 25,
  remove_empty = TRUE)
}
\arguments{
\item{coccur}{The co-occurrence matrix}

\item{dim_size}{The embedding dimension}

\item{iters}{The number of iterations}

\item{remove_empty}{Flag to remove vectors that are identically 0}
}
\value{
PCA embedding
}
\description{
Constructs PCA embedding based raw co-occurrence counts
}
