% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec_fit.R
\name{construct_pmi}
\alias{construct_pmi}
\title{Construct pointwise mutual information matrix}
\usage{
construct_pmi(cooccur, singletons, N, smooth = 0.75)
}
\arguments{
\item{singletons}{The dataframe of CUIs and their counts}

\item{N}{The number of bins}

\item{smooth}{The smoothing factor}

\item{coccur}{The co-occurrence matrix}
}
\value{
A dataframe of pointwise mutual information
}
\description{
Construct pointwise mutual information matrix
}
