% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec_fit.R
\name{construct_sppmi}
\alias{construct_sppmi}
\title{Construct the shifted positive pointwise mutual information matrix}
\usage{
construct_sppmi(pmi, k = 1)
}
\arguments{
\item{pmi}{Pointwise mutual information matrix from \code{\link{construct_pmi}}}

\item{k}{Shift in SSPMI formula}
}
\value{
The shifted positive pointwise mutual information matrix
}
\description{
Construct the shifted positive pointwise mutual information matrix
}
