% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec_fit.R
\name{construct_word2vec_embedding}
\alias{construct_word2vec_embedding}
\title{Construct word2vec embedding}
\usage{
construct_word2vec_embedding(sppmi, dim_size = 100, iters = 25,
  remove_empty = TRUE, use_sum = TRUE)
}
\arguments{
\item{sppmi}{SPPMI matrix from \code{\link{construct_sppmi}}}

\item{dim_size}{The embedding dimension}

\item{iters}{The number of iterations}

\item{remove_empty}{Flag to remove vectors that are identically 0}

\item{use_sum}{Flag to add \eqn{V_d\sqrt{\Sigma_d}} to embedding calculation}
}
\value{
word2vec
}
\description{
Construct word2vec embedding
}
