% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Cosine similarity between two vectors.}
\usage{
cosine_similarity(vec1, vec2)
}
\arguments{
\item{vec1}{First vector}

\item{vec2}{Second vector}
}
\value{
The cosine similarity between two vectors
}
\description{
Cosine similarity between two vectors.
}
