% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cui2vec-package.R
\docType{package}
\name{cui2vec-package}
\alias{cui2vec}
\alias{cui2vec-package}
\title{cui2vec: cui2vec Word Embeddings for Multimodal Data}
\description{
We present a new set of embeddings for medical concepts learned 
    using an extremely large collection of multimodal medical data. 
    Leaning on recent theoretical insights, we demonstrate how an insurance 
    claims database of 60 million members, a collection of 20 million clinical 
    notes, and 1.7 million full text biomedical journal articles can be 
    combined to embed concepts into a common space, resulting in the largest 
    ever set of embeddings for 108,477 medical concepts. To evaluate our 
    approach, we present a new benchmark methodology based on statistical 
    power specifically designed to test embeddings of medical concepts.
}
\author{
\strong{Maintainer}: Andrew Beam \email{andrew_beam@hms.harvard.edu}

Authors:
\itemize{
  \item Benjamin Kompa \email{kompa@fas.harvard.edu}
  \item Allen Schmaltz \email{aschmaltz@hsph.harvard.edu}
}

}
\keyword{internal}
