% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_embeddings}
\alias{load_embeddings}
\title{Load an embedding from a csv file.}
\usage{
load_embeddings(filename, join_semantic_type = TRUE)
}
\arguments{
\item{filename}{File name to be loaded. CUIs should be in first column.}

\item{join_semantic_type}{Flag to join semantic type}
}
\value{
Dataframe containing the embeddings (and semnatic type)
}
\description{
Load an embedding from a csv file.
}
