% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{query_similarity}
\alias{query_similarity}
\title{Similarity between embedding vectors and query vector}
\usage{
query_similarity(word_vectors, query, sort_result = TRUE)
}
\arguments{
\item{word_vectors}{The embedding vectors}

\item{query}{The query vector}

\item{sort_result}{Flag whether to sort by cosine similarity}
}
\value{
List of cosine similarities to query vector for each word in the embedding
}
\description{
Similarity between embedding vectors and query vector
}
