% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_benchmarks.R
\name{run_all_benchmarks}
\alias{run_all_benchmarks}
\title{Run all benchmarks}
\usage{
run_all_benchmarks(embedding_df)
}
\arguments{
\item{embedding_df}{The embedding dataframe with bound semantic types}
}
\value{
List with results of all benchmarks
}
\description{
Run all benchmarks
}
