% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{semantic_type}
\alias{semantic_type}
\title{UMLS Semantic Type Data}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3063795 rows and 3 columns.}
\usage{
data(semantic_type)
}
\description{
Semnatic type data and English descriptions for 3,063,795 CUIs
}
\examples{
data(semantic_type)
}
\keyword{datasets}
